Fixed theming for custom courses
