+++ src/app/components/planner/quarters/courses/CourseCard.tsx
@@ -14,15 +14,7 @@ import { CourseTerm, StoredCourse } from "@customTypes/Course";
 import { Label } from "@customTypes/Label";
 import { DraggableProvided } from "@hello-pangea/dnd";
 import { WarningAmberRounded } from "@mui/icons-material";
-import {
-  Card,
-  CardContent,
-  CssVarsProvider,
-  Grid,
-  Link,
-  Typography,
-  extendTheme,
-} from "@mui/joy";
+import { Card, CardContent, Grid, Link, Typography } from "@mui/joy";
 import { useContext, useState } from "react";
 
 import CourseLabel from "./CourseLabel";
@@ -34,12 +26,6 @@ export interface CourseCardProps {
   provided?: DraggableProvided;
   quarterId?: string;
   customDeleteCourse?: () => void;
-}
-
-declare module "@mui/joy/Card" {
-  interface CardPropsColorOverrides {
-    custom: true;
-  }
 }
 
 export default function CourseCard({
@@ -62,25 +48,6 @@ export default function CourseCard({
     ...draggableStyle,
   });
   const isEnrolledCourse = quarterId !== undefined;
-
-  const theme = extendTheme({
-    components: {
-      JoyCard: {
-        styleOverrides: {
-          root: ({ ownerState, theme }) => ({
-            ...(ownerState.color === "custom" && {
-              invertedColors: true,
-              backgroundColor: "#E6E6FA", //lavender
-              [theme.getColorSchemeSelector("dark")]: {
-                backgroundColor: "#231645",
-              },
-            }),
-          }),
-        },
-      },
-    },
-  });
-
   function handleShowCourseInfoModal(course: StoredCourse) {
     const courseTerm = [
       course,
@@ -100,7 +67,8 @@ export default function CourseCard({
     }
   }
 
-  function handleDeleteCourse(quarterId: string, index: number) {
+  function handleDeleteCourse(quarterId: string | undefined, index: number) {
+    if (!quarterId) return;
     if (customDeleteCourse) {
       customDeleteCourse();
     } else {
@@ -109,63 +77,52 @@ export default function CourseCard({
   }
 
   return (
-    <CssVarsProvider theme={theme}>
-      <Card
-        ref={provided?.innerRef}
-        {...provided?.draggableProps}
-        {...provided?.dragHandleProps}
-        size="sm"
-        variant="soft"
-        color={cardColor()}
-        className="hover:opacity-50"
-        style={{
-          ...getItemStyle(provided?.draggableProps.style),
-          height: "35px",
-          justifyContent: "center",
-        }}
-        onMouseEnter={() => setHighlighted(true)}
-        onMouseLeave={() => setHighlighted(false)}
-      >
-        <CardContent>
-          <Grid
-            container
-            alignItems="center"
-            justifyContent="start"
-            spacing={1}
-          >
-            <Grid xs={10} className="flex flex-row whitespace-nowrap">
-              <Title
-                course={course}
-                onShowCourseInfoModal={handleShowCourseInfoModal}
-                quarterId={quarterId}
-              />
-              <CourseLabelList
-                labels={getCourseLabels(course)}
-                ge={course.ge}
-              />
-            </Grid>
-            <Grid xs={1}>
-              {quarterId !== undefined && (
-                <CloseIconButton
-                  onClick={() => handleDeleteCourse(quarterId, index)}
-                  sx={{
-                    visibility: highlighted ? "visible" : "hidden",
-                  }}
-                />
-              )}
-              {isCustom && (
-                <CloseIconButton
-                  onClick={() => handleRemoveCustom(index)}
-                  sx={{
-                    visibility: highlighted ? "visible" : "hidden",
-                  }}
-                />
-              )}
-            </Grid>
+    <Card
+      ref={provided?.innerRef}
+      {...provided?.draggableProps}
+      {...provided?.dragHandleProps}
+      size="sm"
+      variant="soft"
+      color={cardColor()}
+      className="hover:opacity-50"
+      style={{
+        ...getItemStyle(provided?.draggableProps.style),
+        height: "35px",
+        justifyContent: "center",
+      }}
+      onMouseEnter={() => setHighlighted(true)}
+      onMouseLeave={() => setHighlighted(false)}
+    >
+      <CardContent>
+        <Grid container alignItems="center" justifyContent="start" spacing={1}>
+          <Grid xs={10} className="flex flex-row whitespace-nowrap">
+            <Title
+              course={course}
+              onShowCourseInfoModal={handleShowCourseInfoModal}
+              quarterId={quarterId}
+            />
+            <CourseLabelList labels={getCourseLabels(course)} ge={course.ge} />
           </Grid>
-        </CardContent>
-      </Card>
-    </CssVarsProvider>
+          <Grid xs={1}>
+            {isCustom ? (
+              <CloseIconButton
+                onClick={() => handleRemoveCustom(index)}
+                sx={{
+                  visibility: highlighted ? "visible" : "hidden",
+                }}
+              />
+            ) : (
+              <CloseIconButton
+                onClick={() => handleDeleteCourse(quarterId, index)}
+                sx={{
+                  visibility: highlighted ? "visible" : "hidden",
+                }}
+              />
+            )}
+          </Grid>
+        </Grid>
+      </CardContent>
+    </Card>
   );
 }
